titlebar: refactore view component and layout
