+++ ninchatsdk/src/main/java/com/ninchat/sdk/ninchatqueue/view/NinchatQueueActivity.kt
@@ -11,6 +11,7 @@ import com.ninchat.sdk.ninchatqueue.presenter.INinchatQueuePresenter
 import com.ninchat.sdk.ninchatqueue.presenter.NinchatQueuePresenter
 import com.ninchat.sdk.utils.misc.Parameter
 import kotlinx.android.synthetic.main.activity_ninchat_queue.*
+import kotlinx.android.synthetic.main.activity_ninchat_queue.view.*
 
 class NinchatQueueActivity : NinchatBaseActivity(), INinchatQueuePresenter {
     var queueId: String? = null
@@ -48,7 +49,7 @@ class NinchatQueueActivity : NinchatBaseActivity(), INinchatQueuePresenter {
         ninchatQueuePresenter.updateQueueView(ninchat_queue_activity)
 
         //4: add actions
-        ninchatQueuePresenter.mayBeAttachTitlebar(ninchat_queue_activity, callback = {
+        ninchatQueuePresenter.mayBeAttachTitlebar(ninchat_queue_activity.ninchat_titlebar, callback = {
             this.onClose(ninchat_queue_activity)
         })
 
@@ -76,7 +77,8 @@ class NinchatQueueActivity : NinchatBaseActivity(), INinchatQueuePresenter {
             }
             ninchatQueuePresenter.updateQueueId(intent = data)
             ninchatQueuePresenter.updateQueueView(ninchat_queue_activity)
-            ninchatQueuePresenter.mayBeAttachTitlebar(ninchat_queue_activity, callback = {
+
+            ninchatQueuePresenter.mayBeAttachTitlebar(ninchat_queue_activity.ninchat_titlebar, callback = {
                 this.onClose(ninchat_queue_activity)
             })
         }
@@ -96,7 +98,7 @@ class NinchatQueueActivity : NinchatBaseActivity(), INinchatQueuePresenter {
 
     override fun onQueueUpdate() {
         ninchatQueuePresenter.updateQueueView(ninchat_queue_activity)
-        ninchatQueuePresenter.mayBeAttachTitlebar(ninchat_queue_activity, callback = {
+        ninchatQueuePresenter.mayBeAttachTitlebar(ninchat_queue_activity.ninchat_titlebar, callback = {
             this.onClose(ninchat_queue_activity)
         })
     }
