+++ ninchatsdk/src/main/java/com/ninchat/sdk/titlebar/presenter/NinchatTitlebarPresenter.kt
@@ -1,0 +1,5 @@
+package com.ninchat.sdk.titlebar.presenter
+
+class NinchatTitlebarPresenter {
+
+}
