+++ ninchatsdk/src/main/res/layout/ninchat_titlebar_with_agent_info_and_job_title.xml
@@ -1,0 +1,57 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:gravity="center"
+    android:orientation="horizontal">
+
+    <ImageView
+        android:id="@+id/ninchat_chat_titlebar_avatar_img"
+        android:layout_width="@dimen/ninchat_chat_activity_avatar_size"
+        android:layout_height="@dimen/ninchat_chat_activity_avatar_size"
+        android:layout_marginEnd="@dimen/ninchat_chat_activity_avatar_margin"
+        android:adjustViewBounds="true"
+        android:scaleType="fitCenter"
+        tools:src="@drawable/ninchat_chat_avatar_left" />
+
+
+    <LinearLayout
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_weight="0.8"
+        android:orientation="vertical">
+
+        <TextView
+            android:id="@+id/ninchat_chat_titlebar_agent_name"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginEnd="@dimen/ninchat_chat_activity_text_margin"
+            android:autoLink="none"
+            android:ellipsize="end"
+            android:maxLines="2"
+            android:textColor="@color/ninchat_color_modal_text"
+            android:textStyle="normal|bold"
+            tools:text="Agent name" />
+
+        <TextView
+            android:id="@+id/ninchat_chat_titlebar_agent_jobtitle"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginEnd="@dimen/ninchat_chat_activity_text_margin"
+            android:autoLink="none"
+            android:ellipsize="end"
+            android:maxLines="2"
+            android:textColor="@color/ninchat_color_modal_text"
+            tools:text="Agent job title placeholder" />
+    </LinearLayout>
+
+    <View
+        android:id="@+id/view_padding"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        android:layout_weight="0.2"
+        android:visibility="invisible" />
+
+
+</LinearLayout>
