+++ ninchatsdk/src/main/res/layout/ninchat_titlebar_with_placeholder.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:gravity="center"
+    android:orientation="horizontal">
+
+    <ImageView
+        android:layout_width="@dimen/ninchat_chat_activity_avatar_size"
+        android:layout_height="@dimen/ninchat_chat_activity_avatar_size"
+        android:layout_marginEnd="@dimen/ninchat_chat_activity_avatar_margin"
+        android:adjustViewBounds="true"
+        android:scaleType="fitCenter"
+        android:src="@drawable/ninchat_chat_avatar_placeholder" />
+
+    <TextView
+        style="@style/NinchatTheme.TitlebarPlaceholder"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginEnd="@dimen/ninchat_chat_activity_text_margin"
+        android:layout_weight="0.8"
+        android:autoLink="none"
+        android:ellipsize="end"
+        android:textColor="@color/ninchat_color_modal_text" />
+
+    <View
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        android:layout_weight="0.2"
+        android:visibility="invisible" />
+
+
+</LinearLayout>
