+++ ninchatsdk/src/main/res/values/dimens.xml
@@ -106,7 +106,7 @@
     <dimen name="ninchat_questionnaire_checkbox_margin">10dp</dimen>
 
     <!-- Titlebar -->
-    <dimen name="ninchat_title_bar_height">70dp</dimen>
+    <dimen name="ninchat_title_bar_height">-2dp</dimen>
     <dimen name="ninchat_chat_activity_text_margin">15dp</dimen>
